third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/conv2d.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] conv2d float16 4D both input and filter non-constant tensors default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 0 should be close enough to expected 1.5322265625 by ULP distance: expected a number less than or equal to 18 but got 15905
[FAIL] conv2d float16 4D both input and filter constant tensors default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 0.0078125 should be close enough to expected 1.5322265625 by ULP distance: expected a number less than or equal to 18 but got 7713
[FAIL] conv2d float16 4D input and filter tensors default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 512 should be close enough to expected 1.5322265625 by ULP distance: expected a number less than or equal to 18 but got 8671
[FAIL] conv2d float16 4D input and filter tensors options.padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 512 should be close enough to expected 1.0390625 by ULP distance: expected a number less than or equal to 18 but got 9176
[FAIL] conv2d float16 4D input and filter tensors options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 512 should be close enough to expected 1.5322265625 by ULP distance: expected a number less than or equal to 18 but got 8671
[FAIL] conv2d float16 4D input and filter tensors options.dilations
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -512 should be close enough to expected 1.3603515625 by ULP distance: expected a number less than or equal to 18 but got 41615
[FAIL] depthwise conv2d float16 4D input and filter tensors options.groups= input_channels
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 2 should be close enough to expected 1.1806640625 by ULP distance: expected a number less than or equal to 8 but got 839
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nchw'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nhwc'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.filterLayout='oihw'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.filterLayout='hwio'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.filterLayout='ohwi'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.filterLayout='ihwo'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nhwc' and options.filterLayout='oihw'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nhwc' and options.filterLayout='hwio'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nhwc' and options.filterLayout='ohwi'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors options.inputLayout='nhwc' and options.filterLayout='ihwo'
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -2 should be close enough to expected 0.884765625 by ULP distance: expected a number less than or equal to 8 but got 34028
[FAIL] conv2d float16 4D input and filter tensors 1D options.bias
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -512 should be close enough to expected 1.6982421875 by ULP distance: expected a number less than or equal to 8 but got 41269
[FAIL] conv2d float16 4D input and filter tensors all options
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual -512 should be close enough to expected 0.6591796875 by ULP distance: expected a number less than or equal to 8 but got 42682
[FAIL] conv2d float16 4D input and filter tensors, both negative input tensor and options.bias
  assert_less_than_equal: assert_array_approx_equals_ulp: test conv2d float16 actual 0 should be close enough to expected -0.8271484375 by ULP distance: expected a number less than or equal to 8 but got 15006
Harness: the test ran to completion.
